Move validation check from `HalfEdge` to `Cycle`
`HalfEdge` will soon only reference a single `SurfaceVertex`. Since this
validation check relies on having access to both, it needs to move to
the level of `Cycle`.
